build(webpack): split config for modularity
