+++ cfg/webpack/production/general.js
@@ -1,0 +1,15 @@
+const path = require('path');
+
+
+module.exports = {
+  output: {
+    filename: '[name].[contenthash].js',
+    path: path.resolve(__dirname, 'dist'),
+    clean: true,
+  },
+  optimization: {
+    splitChunks: {
+      chunks: 'all',
+    },
+  },
+};
